+++ connector/utils/build.gradle
@@ -1,0 +1,9 @@
+plugins {
+    id 'blaze-query.java-conventions'
+}
+
+dependencies {
+    api libs.jackson.databind
+}
+
+description = 'blaze-query-connector-utils'
